+++ Chord Provider/PDFBuild/Core/Elements/PDFBuild+Text.swift
@@ -28,6 +28,13 @@ extension PDFBuild {
             )
         }
 
+        /// Init the **text** element
+        /// - Parameters:
+        ///   - text: The text as `NSAttributedString`
+        init(_ text: NSAttributedString) {
+            self.text = text
+        }
+
         /// Draw the **text**
         /// - Parameters:
         ///   - rect: The available rectangle