+++ Chord Provider/Structures/Song/Song+Section+Line+strum.swift
@@ -9,6 +9,13 @@ import Foundation
 
 extension Song.Section.Line {
 
+    struct Strum: Equatable, Codable, Identifiable, Hashable {
+        var id: Int = 0
+        var strum: String = ""
+        var beat: String = ""
+        var tuplet: String = ""
+    }
+
     /// Convert strum characters in the source to fancy symbols
     static var strumCharacterDict: [String: String] {
         [