+++ CHANGELOG.md
@@ -18,3 +18,7 @@
 - Bugfixes for chord editor
 - Add color and font options for viewing the song
 - Add PDF options for size and padding
+
+# 6.x.x (---)
+
+- Improve strum environment